Added SASS; display of strings is working
Added SASS as a dependency to make writing styles easier.  Also wrote a component to handle requesting and displaying strings; while the component works, it seems like the FS2020 developers decided that yielding magic strings for the aircraft make and type is acceptable, so now I'm going to need to figure out how to map those to usable user-facing values.